# OrderService/src/main/resources/api.yaml
swagger: '2.0'
info:
  description: 'API to manage restaurant orders'
  version: 1.0.0
  title: Swagger OrderService
basePath: /v2/api/restaurant
tags:
  - name: order
    description: Restaurant Orders
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /orders:
    get:
      tags:
        - order
      summary: Queries restaurant Orders
      description: 'Queries restaurant Orders'
      operationId: getOrders
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Order'
        '400':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '412':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '503':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
    post:
      tags:
        - order
      summary: Place a restaurant Order
      description: 'Creates a request order to the Restaurant'
      operationId: postOrder
      parameters:
        - in: body
          name: body
          description: order
          required: true
          schema:
            $ref: '#/definitions/OrderData'
      responses:
        '201':
          description: successful operation
          schema:
            $ref: '#/definitions/Order'
        '400':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '412':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '503':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
  /orders/{id}:
    get:
      tags:
        - order
      summary: Queries specific Order
      description: 'Queries specific Order'
      operationId: getOrder
      parameters:
        - in: path
          name: id
          description: orderId
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Order'
        '400':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '412':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '503':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
    put:
      tags:
        - order
      summary: Updates an order
      description: 'Updates an order'
      operationId: putOrder
      parameters:
        - in: path
          name: id
          description: orderId
          required: true
          type: integer
        - in: body
          name: body
          description: order
          required: true
          schema:
            $ref: '#/definitions/Order'
      responses:
        '201':
          description: successful operation
          schema:
            $ref: '#/definitions/Order'
        '400':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '412':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '503':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - order
      summary: deletes an order
      description: 'deletes an order'
      operationId: deleteOrder
      parameters:
        - in: path
          name: id
          description: orderId
          required: true
          type: integer
      responses:
        '204':
          description: successful operation
        '400':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '403':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '404':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '412':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
        '503':
          description: Invalid Order
          schema:
            $ref: '#/definitions/Error'
definitions:
  OrderData:
    type: object
    required:
      - quantity
      - shipDate
    properties:
      quantity:
        type: integer
        format: int32
      shipDate:
        type: string
        format: date-time
  Order:
    required:
      - id
      - status
      - complete
      - quantity
      - shipDate
    properties:
      id:
        type: integer
        format: int64
      status:
        type: string
        description: Order Status
        enum:
          - placed
          - approved
          - delivered
      complete:
        type: boolean
      quantity:
        type: integer
        format: int32
      shipDate:
        type: string
        format: date-time
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: string
      message:
        type: string
externalDocs:
  description: Find out more about Swagger
  url: 'http://swagger.io'